- allow fractional framerates - on download failure go through all links in the list - fix for log entry when deleting oldest sheep - F2 status: added membership information (also updated code for query.php to handle this new information from server) - removed default avi - removed default server handling - fixed location of "No sheep downloaded yet" message - show number of sheep downloaded from server X/Y, where Y is total number of sheep on server - removed power source information from F4 (it is on F3 only) [Windows only] - autodetection of monitor numbers

+++ trunk/client_generic/Client/Player.h
@@ -66,10 +66,10 @@ private:
 	Base::CTimer	m_Timer;
 
 	//	Goal decoding framerate.
-	uint32			m_PlayerFps;
+	fp8			m_PlayerFps;
 
 	//	Goal display framerate;
-	uint32			m_DisplayFps;
+	fp8			m_DisplayFps;
 
 	//	Fullscreen or not.
 	bool			m_bFullscreen;
@@ -138,7 +138,7 @@ private:
 #endif
 
 			inline void		PlayCountsInitOff()					{	m_InitPlayCounts = false; };
-			inline void		Framerate( const uint32 _fps )		{	m_PlayerFps = _fps;	};
+			inline void		Framerate( const fp8 _fps )			{	m_PlayerFps = _fps;	};
 			inline void		Fullscreen( const bool _bState )	{	m_bFullscreen = _bState; };
 			inline bool		Stopped()							{	return !m_bStarted;	};
